cybozu: API changes for closing file descriptors early
Previously, the file descriptor in a `cybozu::resource` remains open
throughout its lifetime, so it was unnecessary to protect the reference
to such a file descriptor.

The new design is to allow the reactor to close resource file
descriptors early before the resource is destructed so that yrmcdsd
will be less likely to cause an out of file descriptors error.

For this, we have to proctect a resource file descriptor not to
be closed while it is being used. The new API design implements this
with the following changes:

- `resource::m_fd` is now a private member, so no direct reference is allowed.
- Subclasses should use `resource::with_fd` to get the file descriptor along with a reader lock.
- The reactor thread closes the file descriptor only if it can get a writer lock.